move generic helper functions and resources out of map_generation module
