+++ sketch_map_tool/helpers.py
@@ -1,6 +1,22 @@
 from pathlib import Path
 
+from reportlab.graphics.shapes import Drawing
+
 
 def get_project_root() -> Path:
     """Get root of the Python project."""
     return Path(__file__).resolve().parent.parent.resolve()
+
+
+def resize_rlg_by_width(d: Drawing, size: float) -> Drawing:
+    factor = size / d.width
+    d.scale(factor, factor)
+    d.asDrawing(d.width * factor, d.height * factor)
+    return d
+
+
+def resize_rlg_by_height(d: Drawing, size: float) -> Drawing:
+    factor = size / d.height
+    d.scale(factor, factor)
+    d.asDrawing(d.width * factor, d.height * factor)
+    return d
